Conflicts:
	projects/XTK/src/xtk/cl_XTK_Side_Cluster.cpp

+++ share/scripts/create_shared_object.sh
@@ -121,6 +121,8 @@ mv "$MORISROOT/projects/mains/input_file.cpp" "$TMPDIR/."
 
 ln -s "$workdir/$cppfile.cpp" "$MORISROOT/projects/mains/input_file.cpp"
 
+rm -f "MORISROOT/$builddir/lib/input_file.so"
+
 touch "$MORISROOT/projects/mains/input_file.cpp"
 
 cd "$MORISROOT/$builddir"
@@ -129,7 +131,13 @@ make shared_object_file
 
 cd "$workdir"
 
-mv "$MORISROOT/$builddir/lib/input_file.so" $cppfile.so
+if [ -f "$MORISROOT/$builddir/lib/input_file.so" ];then
+    mv "$MORISROOT/$builddir/lib/input_file.so" $cppfile.so
+else
+    echo ""
+    echo " .so file does not exist due to compiler error"
+    echo ""
+fi
 
 if [ "$dbgflag" = '0' ];then
     rm "$MORISROOT/projects/mains/input_file.cpp"
